+++ packages/react-native/tests/AsyncLink.spec.tsx
@@ -225,6 +225,32 @@ describe("<AsyncLink>", () => {
       const anchor = tree.root.findByType(TouchableHighlight);
       expect(anchor.props.style).toMatchObject(style);
     });
+
+    it('does not overwrite "native" props set on the rendered element', () => {
+      const onPress = jest.fn();
+
+      const routes = prepareRoutes([
+        { name: "Test", path: "" },
+        { name: "Catch All", path: "(.*)" }
+      ]);
+      const router = createRouter(inMemory, routes, {
+        history: {
+          locations: ["/the-initial-location"]
+        }
+      });
+      const Router = createRouterComponent(router);
+
+      const tree = renderer.create(
+        <Router>
+          <AsyncLink name="Test" forward={{ onPress }}>
+            {navigating => <Text>{navigating}</Text>}
+          </AsyncLink>
+        </Router>
+      );
+
+      const anchor = tree.root.findByType(TouchableHighlight);
+      expect(anchor.props.onPress).not.toBe(onPress);
+    });
   });
 
   describe("ref", () => {