+++ packages/react-native/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Spread `forward` props to link elements before "native" props.
+
 ## 2.0.0-beta.6
 
 * Revert snake case. `create_router_component` is now `createRouterComponent`.